Renamed class GameDataManger to GameDataController Renamed folder GameDataManager to GameData
